Oauth

+++ routers/internshipRouter.js
@@ -1,6 +1,8 @@
 const express = require("express");
 const prisma = require("../prisma/prisma");
 const { catchAsync } = require("../utils/catchAsync");
+const { isLoggedIn } = require("../middleware/isLoggedIn");
+const { checkRole } = require("../middleware/checkRole");
 
 const router = express.Router();
 
@@ -13,9 +15,18 @@ router
     })
   )
   .post(
+    isLoggedIn,
+    checkRole("company"),
     catchAsync(async (req, res) => {
+      const company = await prisma.company.findUnique({
+        where: {
+          contactEmail: req.user.profile.emails[0].value,
+        },
+      });
+
       const newListing = await prisma.internship_listing.create({
         data: {
+          companyID: Number(company.companyID),
           ...req.body,
           ...(req.body.startDate
             ? { startDate: new Date(req.body.startDate) }
@@ -41,10 +52,29 @@ router
     })
   )
   .patch(
+    isLoggedIn,
+    checkRole("company"),
     catchAsync(async (req, res) => {
+      const company = await prisma.company.findUnique({
+        where: {
+          contactEmail: req.user.profile.emails[0].value,
+        },
+      });
+
+      const listing = await prisma.internship_listing.findUnique({
+        where: {
+          companyID: Number(company.companyID),
+        },
+      });
+
+      if (!listing) {
+        return res.status(401).json({ error: "Unauthorized" });
+      }
+
       const updatedListing = await prisma.internship_listing.update({
         where: {
           listingID: Number(req.params.id),
+          companyID: Number(company.companyID),
         },
         data: {
           ...req.body,
@@ -59,10 +89,29 @@ router
     })
   )
   .delete(
+    isLoggedIn,
+    checkRole("company"),
     catchAsync(async (req, res) => {
+      const company = await prisma.company.findUnique({
+        where: {
+          contactEmail: req.user.profile.emails[0].value,
+        },
+      });
+
+      const listing = await prisma.internship_listing.findUnique({
+        where: {
+          companyID: Number(company.companyID),
+        },
+      });
+
+      if (!listing) {
+        return res.status(401).json({ error: "Unauthorized" });
+      }
+
       const deletedListing = await prisma.internship_listing.delete({
         where: {
           listingID: Number(req.params.id),
+          companyID: Number(company.companyID),
         },
       });
